first iteration of autoencoders

+++ src/utils/dataset_utils.py
@@ -3,4 +3,4 @@ from torch import Tensor, from_numpy
 
 
 def read_npy(path) -> Tensor:
-    return from_numpy(np.load(path))
+    return from_numpy(np.load(path)).float()
